+++ src/main/java/com/example/servicehub/web/controller/ServiceController.java
@@ -1,4 +1,4 @@
-package com.example.servicehub.web;
+package com.example.servicehub.web.controller;
 
 import com.example.servicehub.dto.*;
 import com.example.servicehub.security.authentication.ClientPrincipal;
